+++ CHANGELOG.md
@@ -7,7 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- [SQL] Support for specifying connector properties in SQL
+  ([#1896](https://github.com/feldera/feldera/pull/1896))
 - [SQL] Support for ARRAY_AGG aggregation function.
+  ([#1862](https://github.com/feldera/feldera/pull/1862))
 - [SQL] Support for hopping windows using table functions
   ([#1855](https://github.com/feldera/feldera/pull/1855))
 - API: Support skipping the serialization of schema ID in Avro format.